+++ report-framework/src/components/feedback/index.js
@@ -1,3 +1,4 @@
 export { EmptyState } from './EmptyState'
 export { LoadingState } from './LoadingState'
 export { ErrorState } from './ErrorState'
+export { ErrorBoundary, withErrorBoundary } from './ErrorBoundary'